Move '_forced' aapt cli flag handling up to aaptTask
This pushes flag handling into the tools that it effects.
Instead of repeasting this code every time we render agrs, now it gets
handled by the task tha owns the flag.

Added TODOs to map output streams to workunits when executing tools.
Killed the ['apk'] product as there is nothing downstream consuming it yet.