Move Android key definitions into 3rdparty
This framework adds a dir for release BUILD files and keys. Even
though release builds are not supported, I believe we are better
providing a safe framework than letting users fend for themselves.

Without explicitely showing a safe place for the releasekeys,
we are risking someone trying to figure it out on their own
and accidentally exposing a production key in a vcs.

So, this is not perfect, but much better than nothing.