JarsignerTask and Keystore Target
This adds a new Keystore target and a JarsignerTask. It outputs
the finished apk, which can be verified by pointing it at the sample
code and installing the apk it outputs.

Pants can now output debug and publishable release builds.
The release keystore is removed from version control,
although we will be layering another level of protections
with the release keystore handling in the future.

This bundles an (insecure) debug.keystore, which is scheduled
to be removed soon once we add in the KeytoolTask.

Testing Done:
CI.sh passes.
Travis passes [0]

[0] https://travis-ci.org/pantsbuild/pants/jobs/32476705

Reviewed at https://rbcommons.com/s/twitter/r/883/