Properly order resources for pants goal test and pants goal run.
Fixed a problem with goal run and test for resources being put on the classpath in the
wrong order when multiple resource sources define a file with different contents
at the same resource path.

Added a project with resources that collide and an integration test to prove
that resources are bundled and added to the classpath in the right order.

Note that the JVM tests will still fail if all run together with a command like:

   pants goal test maven_src/resource_collision::

because pants runs all the tests in the same junit test runner instance.

Testing Done:
Added an integration test for pants goal bundle and pants goal run.

Ci is baking at https://travis-ci.org/pantsbuild/pants/builds/32999609

Bugs closed: 449

Reviewed at https://rbcommons.com/s/twitter/r/845/