docs: add pages w/not much content (yet), howto_plugin, howto_customize
howto_plugin: how to write a Pants plugin

howto_customize: how to write a Pants customized for your org

So far, neither of these pages has much content. But I hope we'll
be better about capturing some of this info if we have a place to put it.

Reviewed at https://rbcommons.com/s/twitter/r/938/