update and narrow down our "setting up Pants is hard" message
pantsbuild.pants is is a thing. So stop saying it's not installable.

We had commented out stuff about how to install it. Un-comment it.

For a venv-y example, see twitter/commons; not pants' repo, that was only true for a brief time.

BUT, setting up Pants in your code workspace still requires a PhD in Pants-ology. So shift our scary warning over to the setup_repo page.

Reviewed at https://rbcommons.com/s/twitter/r/1007/