Plumb new-style options into the Context.
Allows code to reference option values the new way.

Renamed the overloaded self.options in Command to self.old_options/self.new_options,
to avoid confusion. Previously these collided and the only reason it worked was that
self.options happened to be reset to the right one just before using it.

Change a single flag reference to use the new path, as proof that it works.

With this change we can not only register flags the new way but also reference the option
values the new way. Once everything is ported to use these, we can switch the parser to
the new one, and the transition will be complete.

Testing Done:
ci baking.

Reviewed at https://rbcommons.com/s/twitter/r/1102/