Factor in the jvm version string into the nailgun executor fingerprint
In some configurations, the path to the 'java' executable stays the same but can be manipulated to a different version via a symlink or environment variable.
This causes a mismatch between the already running nailgun server and the jvm that Pants thinks it has chosen.

Testing Done:
Manual testing to make sure that ng servier is killed if symlink to java executable is changed.

Writing a unit test for this looked prohibitivly difficult.

Bugs closed: 639

Reviewed at https://rbcommons.com/s/twitter/r/1122/