Handle list-typed options (action='append') correctly.
Ensure that we read them from config in the right way, and apply type
conversion to the list members.

Added test cases for this.

Testing Done:
Ran the options unittests before and after the fix, to verify
that the test discriminates this.

CI passed.

Reviewed at https://rbcommons.com/s/twitter/r/1183/