jvmdoc_gen: In the event of an exception, call get() on the remaining futures.
jvmdoc_gen: In the event of an exception, we want to call get() on the remaining futures because otherwise we get errors on exit when multiprocessing tries to kill them because what is dead may never die.

Ideally, we would kill them instead of waiting for them, but I don't see how to do that.

Testing Done:
./pants goal test tests/python/pants_test/tasks:jvmdoc_gen

It would be insanely difficult to write a real test for this, because the failure happens in an atexit handler.

Reviewed at https://rbcommons.com/s/twitter/r/1202/