Allow more succinct invocations with passthru args.
If passthru args "belong" to a goal scope then all tasks in that goal get them.

E.g., you can do:

pants goal run <python-target> -- passthru args

and the run.py task will get the passthru args. So will run.jvm, but that's OK
because it's a no-op for <python-target>.

This obviously is only good when in practice only one passthru-arg-supporting
task in that goal is not a no-op. However that is the common case.

Reviewed at https://rbcommons.com/s/twitter/r/1323/