doc: fix JVMProjects problems I didn't want to think about in rb/1176
Don't claim 'goal jar' is one of "The Usual Commands". It's kind of obscure and scary.

An extra "is" was in code font. Yeah, "is" is valid Python. But this wasn't Python, this was English.

Updated ivy mention: use new flags/config.

Tweaked nailgun description.

While I was in there: removed a section on getting help. It was redundant with
first_tutorial, except that it hadn't kept up to date :-(

Reviewed at https://rbcommons.com/s/twitter/r/1358/