Fixup BuildGraph to handle mixes of synthetic and BUILD targets.
An upcoming change will inject synthetic JarLibraries that jvm
tools will depend on.  This tickles a corner where the BuildGraph
can be asked to resolve a SyntheticAddress that has been injected.

Tighten up interfaces removing unused methods and prefixing private
helpers with _ to clarify APIs and tighten them.

Modify tests to validate the new translation of SyntheticAddress to
BuildFileAddress when lookups are made and resurrect tests in
BuildGraphTest.

This also fixes some minor fallout in production code and tests.