Make AntlrGen an out-of-the-box task.
This removes the need for adding pants.ini or BUILD.tools
entries to get AntlrGen working.  This change is limited in
scope to AntlrGen, but the path is paved for antlr-gen's
custom fixed-dep injection in the absence of config to be lifted
to a jvm tool helper so that all jvm tool taks can easily be
converted to pants.ini/BUILD.tools free Tasks that "just work".