Allow configuring goal aliases
Usually unrecognized args (eg not scopes or -foo) become target specs.
If pants.ini configures goal_aliases, look for specs that match aliases
and if found remove them from specs and add their configured goal names
to the goals.

This might be prettier if we instead told the arg splitter about aliases,
by including them in known-scopes, rather than rescuing them from specs
but this is good enough for now, and that approach has potential to muck
with other args, as it adds the aliases as a scope, which affects more than
just which list they themselves get sorted into.