allow adding to pythonpath via pant.ini
Looks like this used to work, based on a comment (recently removed) that said:
```
"""
  To add additional paths to sys.path, add a block to the config similar to the following:
  roots: ['src/python/pants_internal/test/',]
"""
```
The provides an alternative to `PYTOHNPATH=custom/pants/stuff ./pants goal ...` when using
an in-repo custom register.py.

Note that this is unrelated to plugins distributed as dists and installed in the (virtual)env
from which pants is run / loaded with pkg_resources, or source backends packaged into a custom
dist or pex of pants.

This change restores that functionality, though it is now called `python-path` for clarity and
lives in the `[backends]` config section.

Bugs closed: 856

Reviewed at https://rbcommons.com/s/twitter/r/1457/