Cache ivy report dependency traversals consistently
For depmap calls using project-info, ~30% of the time was taken by retraversing the dependency graph from the ivy report xml. This encapsulates memoizing and cycle handling into a single method, and replaces the memoization / cycle handling done outside ivy utils with calls to the new method. It also adds a few tests for reports with cycles and diamond dependencies