Optionally include direct or transitive dependees of changed targets
Frequently the goal of asking for changed targets is to determine what needs to be built or tested to ensure some change is valid.

Dependees of a changed target often need to be built or tested in addtion to the target itself to ensure a given change hasn't broken them, eg if the changes to the target alter an interface.

Frequently just checking direct dependees is "good enough" and faster than all transitive dependees,
so both are provided as options.

Testing Done:
Added tests to test_what_changed to run with --include-dependees=direct and --include-dependees=transitive.

https://travis-ci.org/pantsbuild/pants/builds/46022156

Bugs closed: 905

Reviewed at https://rbcommons.com/s/twitter/r/1567/