Refactor setting of python-path in pants.ini
We recently added the ability to set pythonpath in pants.ini.
This change was made to document the setting.  Unfortunately, the previous
name was hyphenated, and the new options system would translate that into an
underscore automatically.  I removed the separator to simplify this, plus
it should be easier to remember as PYTHONPATH in the environment doesn't
have a separator anyway.