Fix topological ordering + chunking bug in jvm_compile.
A recent refactor to removing the topological ordering assumed that callers of `Task.invalidated` did not need their versioned target sets to be returned in topological order.  This is however untrue for `jvm_compile`, which does custom chunking on the VTS and assumes the results are in topological order.

For now I'm plumbing through this option, but we might want to kill this option and force tasks that want topologically ordered targets to call `sort_targets` themselves.

We really should have an integration test for chunked scala compiles using various partitioning strategies.  For now, this break is so bad (the result is just code that somewhat mysteriously doesn't compile, it isn't a pants error) that I wanted to upstream the fix that I've confirmed works internally on our codebase.

Testing Done:
CI is green: https://travis-ci.org/pantsbuild/pants/builds/46780553

Reviewed at https://rbcommons.com/s/twitter/r/1598/