Fix spurious Products requirements.
The trio or require, isrequired and get - the original members of
products - are tricky and they have invalid semantics in the
product-driven RoundEngine.

This change fixes things such that a `.get(<typename>)` does not have
the effect of causing a subsequent `.isrequired(<typename>)` check to be
truthy.

Going forward, the idea of requiring a product for a subset of targets -
the role of the `.require` predicate today - will be subsumed by the
tuple-based engine.