Fix spurious Products requirements.
The trio of require, isrequired and get - the original members of
products - are tricky and they have invalid semantics in the
product-driven RoundEngine.

This change fixes things such that a `.get(<typename>)` does not have
the effect of causing a subsequent `.isrequired(<typename>)` check to be
truthy.

Going forward, the idea of requiring a product for a subset of targets -
the role of the `.require` predicate today - will be subsumed by the
tuple-based engine.

Testing Done:
I ran into this in the ProtobufGen imports handling in a pretty subtle
way when working on an engine change.

In (not so) short, ApacheThriftGen runs 1st, and in the post-processing
phase in the CodeGen superclass where synthetic targets are linked, a
call to `products.get('python')` ends up making 'python' products
required. This in turn influences the subsequent ProtobufGen (another
CodeGen sublclass) run, incorrectly telling it that 'python' protobuf
gen products are required.

CI went green here:
https://travis-ci.org/pantsbuild/pants/builds/48218578

Bugs closed: 989

Reviewed at https://rbcommons.com/s/twitter/r/1662/