Improve RoundEngine lifecycle.
This change lifts all engine planning methods needed to prepare the task
graph for a run to be task classmethods.  In addtion it introduces a new
alternate_target_roots lifecycle method to support more disciplined
replacement of the active target set.  As part of this move,
Context.replace_targets is made ~private to discourage its direct use.

The task graph preparation lifecycle is now the following sequence of
task classmethods:
1. product_types - declare the product types your task is capable of
   producing.
2. register_options - declare options configurable via cmd-line flag
   or config file.
3. alternate_target_roots - propose a different set of target roots to
   use than those specified via the CLI for the active pants run.
4. prepare - request any products needed from other tasks.

The task execution lifecycle is now just construct and then execute,
back to back.

Note that this fix is just a temporary aid in the transition to the tuple
engine.

Fixes along the way:
+ Fixup all Tasks to use the new lifecycle.
+ Fixup tests using Context.replace_targets when they need not have been.
+ Re-work ChangedFileTaskMixin for use with alternate_target_roots in the changed goals.
+ Kill broken and unused BuildLint task.

Several more fixes pointed at by broken tests:
+ Have Depmap use prepare to create a conditional dependency on resolve
  and cleanup DepmapIntegrationTest.
+ Fix ProtobufIntegrationTest.test_source_ordering to ensure its
  assumption of a fresh gen, also fixup tests for "output contains".
+ Fix WireIntegrationTest.test_good to ensure its assumption of a fresh
  gen, also fixup tests for "output contains".