Change Keystore from a target to an object, signed by SignApkTask.
This change changes the Keystore from being a first-class target into
a regular object. Since pants does not actually build a keystore, this
is a better abstraction.

Instead of passing Keystore definitions in BUILD files, now there
is a keystore_config.ini which is passed as an option. This
option is parsed through the usual channels and is not a
threat to be accidentally added to version control.

The config file is in .ini syntax, to lower the barrier to entry
for casual or new users. Upon first build of an android target,
a default config file is put in ~/.pants.d/android. This is done
by the new AndroidConfigUtils class and points to a well-known
keystore installed by the Android SDK.

Design Doc:
https://docs.google.com/document/d/1bhq4L4GW0Q-KMtIBy_AkEg4oWIX6pC5K-XUdMf2fVcg

Tests and integration tests for all new classes are included.