Change SignApk get_apk method to return a list of product paths.
As it was, the method returned a single path. This was fine in
reality, but made for confusing code because it was intentionally
short-circuiting a for loop w/o explanation.

There is no current circumstance where this method will return
a list of more than one item, but better to be clear about
intentions.