Add an env var for enabling the profiler
This adds PANTS_PROFILE, which if exported, will add the cProfiler
module to the python command run to execute pants.  It should be set
to the file you want the profile results to be dumped to. eg

    PANTS_PROFILE=results ./pants goal list ::

You can then pass the pstats formatted file into an analyzer, or a
graph generator to see where that run spent its time.

Reviewed at https://rbcommons.com/s/twitter/r/1305/