docs: Show correct pictures for intellij.html
I also removed a stray 'goal goals' reference. This
required replacing one of the pictures. I am using
a newer version on Intellij than jsirois as well as a
different theme. I am open to the argument that the
aesthetic dissonance is not worth it and it would be better to
leave the old picture in.