Introduce isort to check *.py import ordering.
This change adds build-support/bin/isort.sh and incorporates it in
build-support/bin/ci.sh. By default the script runs in check mode and
exits non-zero if there are one or more files with non-conformant import
order. The script also has a fix mode that will fixup files not in
compliance. This mode was run in this change to normalize the repo.