Bump sbt version to sbt 0.13.6.
Sbt 0.13.6 turned on name hashing by default. This breaks a few
zinc operations: merge, rebase, reload and split. It's reflected
by failure of scriptit tests corresponding to those operations.

We disable those tests until underlying issue is fixed. The problem
is being tracked in #68.

The workaround for users is to disable name hashing. This is not
possible at the moment but a command line option will be introduced
in a subsequent commit.