Fixup ci isort check.
The advice to run the isort in fix mode was not properly escaped and the
fix command would actually run inside backticks - this is now fixed.

Additionally moved the formatting checks to run 1st in ci in their own
section.  These checks are fast and folks will want to know about and
correct these errors before running any other part of ci.