Fixup ivy report generation and opening.
Previously pants would open a browser tab for each ivy configuration in
play.  When doing an ivy open in the same pants run as an idea project
gen this would open 3 browser tabs, one for default (compile), source
and javadoc artifacts.  This change fixes open to only open one browser
tab since the ivy report opened itself has internal tabs that can be
used to navigate to the report for each conf.

It turns out the ivy report internal tab links were also broken.  They
linked to the right file with the wrong extension (xml instead of html).
It turns out the extension is an ivy-report.xsl paramater and so this
change introduces overriding that parameter from the command line used
to run xalan in order to get the links working correctly.