Add SignApk products and Zipalign task.
Zipalign is the final step in the basic Android app build.
This step is only required on apks signed with release keystores.
SignApk now outputs two products, 'debug_apk' and 'release_apk'.
Zipalign requires them both so as to bundle their tasks into a
'binary' goal. But Zipalign only acts upon the 'release_apk'.

SignApk has to put any output apk products into context. As it
stands, the keystore objects and AndroidBinary targets have no
concept of each other. Keystore objects only exist in the
scope of the target's SignApk execute phase. In order to maintain
that, the SignApk has to check for the products on disk.
If it finds a product, then it adds that product to the context.

I also changed the return type of the KeystoreResolver into a dict.
Since it will usually return more than one keystore, it was easier
to test and use as a dictionary.